Fixed a bug in Xorshift random number generators variable size declaration which caused a huge performance issue

+++ avr-gcc/share.c
@@ -1,9 +1,9 @@
 #include "share.h"
 
-static uint64_t x=123456789, y=362436069, z=521288629;
+static uint32_t x=123456789, y=362436069, z=521288629;
 
-uint64_t xorshf96(void) {
-  uint64_t t;
+uint32_t xorshf96(void) {
+  uint32_t t;
   x ^= x << 16;
   x ^= x >> 5;
   x ^= x << 1;
